move util functions out of form
